requirements.txt
flask
flask_sqlalchemy
flask_login
flask_wtf
cryptography
python-dotenv
